feat(dataset): make the filter method extendible to potentially other types of series
The change was needed to make sure that subclassed timeseries having attributes other than "sensor" can also be filtered. It is now done through kwargs.

fix(Dataset.filter): change the names of the parameters to correspond with the BaseTimeseries attrs.

docs: bring the docs notebook to be in line with the changes in Dataset.

test: Better test Dataset